[Yachna]-Edit the existing contact person using their name

+++ Address/Contact.cs
@@ -13,7 +13,7 @@ namespace Address
         public string Address { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public long Zip { get; set; }
+        public int Zip { get; set; }
         public long Phonenumber { get; set; }
         public string Email { get; set; }
     }
